use common mergeData, move ConsultationList to common, delete info from json
