+++ components/common/ConsultationsList/ConsultationsList.tsx
@@ -1,10 +1,12 @@
 import { performRequest } from '@/lib/datocms';
 import { consultancyPlatesQuery } from '@/lib/queries/consultancyPlatesQuery';
 
+import { ConsultationsCardProps } from '@/components/common/ConsultationsCard/type';
+
 import { ConsultationsCard } from '@/components/common/ConsultationsCard';
 
 import consultations from '@/data/consultations.json';
-import { mergeConsultationsData } from '@/utils/mergeData';
+import { mergeData } from '@/utils/mergeData';
 
 export const ConsultationsList = async () => {
   const { consultationsList } = consultations;
@@ -13,18 +15,26 @@ export const ConsultationsList = async () => {
   const plates = data?.consultancyPlate.plates;
 
   const dataToShow = plates
-    ? mergeConsultationsData(plates, consultationsList)
+    ? mergeData(plates, consultationsList)
     : consultationsList;
 
   return (
     <ul className="flex flex-col gap-4 xl:flex-row xl:gap-8">
-      {dataToShow.map(item => {
-        const { title, textCard, format, frequency, duration, linkText, href } =
-          item;
+      {dataToShow.map((item: ConsultationsCardProps) => {
+        const {
+          id,
+          text,
+          textCard,
+          format,
+          frequency,
+          duration,
+          linkText,
+          href,
+        } = item;
         return (
-          <li key={item.id}>
+          <li key={id}>
             <ConsultationsCard
-              title={title}
+              text={text}
               textCard={textCard}
               format={format}
               frequency={frequency}
