+++ data/consultations.json
@@ -4,52 +4,34 @@
   "sectionText": "Під час консультацій я проводжу глибокий внутрішній аналіз, застосувую різноманітні психологічні техніки від класичного психоаналізу до гіпнотерапії; підбираю найкращі методики для швидкого та ефективного розв'язання проблеми.",
   "consultationsList": [
     {
-      "id": 1,
-      "title": "Індивідуальна консультація",
+      "id": "1",
       "textCard": "Особиста підтримка та розвиток через відкрите спілкування та розуміння власних емоцій і думок.",
       "linkText": "Детальніше",
       "href": "/individual-consultation",
       "format": "Онлайн",
       "frequency": "1 раз/тиждень",
       "duration": "50-60 хв/сеанс",
-      "text": "Індивідуальна консультація",
-      "info": [
-        { "id": 1, "infoText": "Онлайн" },
-        { "id": 2, "infoText": "1 раз/тиждень" },
-        { "id": 3, "infoText": "50-60 хв/сеанс" }
-      ]
+      "text": "Індивідуальна консультація"
     },
     {
-      "id": 2,
-      "title": "Сімейна консультація",
+      "id": "2",
       "textCard": "Підтримка гармонії та злагоди в сімейних стосунках шляхом спільного розуміння та вирішення проблем.",
       "linkText": "Детальніше",
       "href": "/family-consultation",
       "format": "Онлайн",
       "frequency": "1 раз/тиждень",
       "duration": "60-90 хв/сеанс",
-      "text": "Сімейна консультація",
-      "info": [
-        { "id": 1, "infoText": "Онлайн" },
-        { "id": 2, "infoText": "1 раз/тиждень" },
-        { "id": 3, "infoText": "60-90 хв/сеанс" }
-      ]
+      "text": "Сімейна консультація"
     },
     {
-      "id": 3,
-      "title": "Групова консультація",
+      "id": "3",
       "textCard": "Підтримка у вирішенні проблем та розвитку особистості через спільний досвід та взаємну підтримку в групі.",
       "linkText": "Детальніше",
       "href": "/group-consultation",
       "format": "Онлайн",
       "frequency": "1 раз/тиждень",
       "duration": "60-120 хв/сеанс",
-      "text": "Групова консультація",
-      "info": [
-        { "id": 1, "infoText": "Онлайн" },
-        { "id": 2, "infoText": "1 раз/тиждень" },
-        { "id": 3, "infoText": "60-120 хв/сеанс" }
-      ]
+      "text": "Групова консультація"
     }
   ]
 }
