+++ utils/mergeData.ts
@@ -1,28 +1,10 @@
-import { UniquenessCardProps } from '@/components/common/UniquenessCard/type';
-import { Plate, StaticData } from '@/components/common/ConsultationsCard/type';
-
-export const mergeData = (
-  datoData: UniquenessCardProps[],
-  staticData: UniquenessCardProps[],
-) => {
-  const result = staticData.map(staticItem =>
+export const mergeData = (datoData: any, staticData: any) => {
+  const result = staticData.map((staticItem: any) =>
     Object.assign(
       staticItem,
-      datoData.find(datoItem => datoItem.text === staticItem.text),
+      datoData.find((datoItem: any) => datoItem.text === staticItem.text),
     ),
   );
-  return result;
-};
 
-export const mergeConsultationsData = (
-  datoData: Plate[],
-  staticData: StaticData[],
-) => {
-  const result = staticData.map(staticItem =>
-    Object.assign(
-      staticItem,
-      datoData.find(datoItem => datoItem.text === staticItem.title),
-    ),
-  );
   return result;
 };
